chore: cleanup and yagni
